requirements.txt
replicate
discord.py
lxml
pybase64
beautifulsoup4
